requirements.txt
fastapi
uvicorn
easyocr
Pillow
numpy
requests
python-dateutil
rapidfuzz
